# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RxDx Coding API",
    "version": "1.1.0",
    "summary": "Turns clinical free text into ICD-10-AM codes, with the evidence for each one.",
    "description": "Every code comes back with the phrase that earned it, the sentence it sits in, a confidence figure and the reasons behind that figure. The service also returns what the note mentions but the engine deliberately refused to code, so a reviewer can see that a denied diagnosis was read and set aside rather than missed.\n\nThis API runs the same engine as the RxDx browser tool — literally the same script file, not a re-implementation — so the two can never give different answers. A parity test fails the build if they ever do.\n\nMeasured on a 60-note hand-labelled corpus: precision 100.0%, recall 98.7%, F1 99.3% at 3-character ICD category; 97.3%, 94.8% and 96.1% at full code. Negation: 123 of 123 traps held.\n\nPRIVACY. Clinical text is held in memory for the life of one request. It is never written to disk, never logged and never forwarded. The access log records a request id, a character count, how many codes came back and how long it took. If the deployment you are calling is the public demonstration instance, do not send real patient text to it — the note still crosses the public internet to reach the server.",
    "contact": { "name": "Mohammed Alshehri", "email": "m7md.alshehri68@gmail.com", "url": "https://www.linkedin.com/in/mh-sh68/" }
  },
  "servers": [{ "url": "/", "description": "this deployment" }],
...
  "paths": {
...
    "/v1/encounter": {
      "post": {
        "tags": ["documentation"],
        "summary": "What still has to be written for this visit",
        "description": "Give the presenting complaint and, optionally, the note as it stands. The service returns the protocol prompts for that complaint, the pre-authorisation requirements the selected payer applies to it, and — the part that matters — what a payer will want that is not in the note yet.\n\nRules triggered by the presenting complaint are this visit's business. Rules triggered only by a condition mentioned in the background come back separately under otherConditionsOnFile, because a doctor documenting chest pain should not be handed the paperwork for a glucometer. That separation is the difference between a tool a clinician uses and one they close.\n\nNothing here is generated. The prompts come from the protocol library and the requirements from the published pre-authorisation protocols of three Saudi insurers.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EncounterRequest" },
              "example": {
                "complaint": "chest pain",
                "payer": "taw",
                "note": "61 y male with central chest pain for two hours radiating to the left arm. ECG shows ST depression. Troponin sent."
              }
            }
          }
        },
        "responses": {
          "200": { "description": "the questions still open", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EncounterResponse" } } } },
          "400": { "description": "no complaint was supplied", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "422": { "description": "the complaint is not in the protocol library", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/v1/code": {
      "post": {
        "tags": ["coding"],
        "summary": "Code one clinical note",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CodeRequest" },
              "example": {
                "text": "45 year old man admitted with community acquired pneumonia. Developed acute kidney injury on day 3. Serial troponins negative, no myocardial infarction.",
                "sex": "male"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "coded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CodeResponse" } } } },
          "400": { "description": "the request could not be read", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "a key is required and was not supplied", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "413": { "description": "the note is larger than this service accepts", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "rate limited", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/v1/code/batch": {
      "post": {
        "tags": ["coding"],
        "summary": "Code many notes in one call",
        "description": "For a coder workstation working through a queue, or a monthly documentation-gap report over past notes. Each answer carries back the id you sent, so nothing is mixed up.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "example": {
            "notes": [
              { "id": "ENC-1001", "text": "Type 2 diabetes with diabetic foot ulcer. Wound swab grew MRSA.", "sex": "female" },
              { "id": "ENC-1002", "text": "Chest pain, non cardiac. Troponin negative." }
            ]
          } } }
        },
        "responses": { "200": { "description": "coded" }, "413": { "description": "too many notes" } }
      }
    },
    "/v1/code-note": {
      "post": {
        "tags": [
          "code-note"
        ],
        "summary": "Code one note: principal and additional diagnoses, evidence, necessity, validation",
        "description": "The RxDx engine is the first and authoritative pass. Every code exists in the ICD-10-AM table at request time; every quote is verified at its offsets; the validation layer runs on every response. The note is held in memory for this request only.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeNoteRequest"
              },
              "example": {
                "clinical_note": "55 y male with central chest pain. Known hypertension. ECG: ST elevation. Troponin elevated. Impression: inferior STEMI. Aspirin and clopidogrel given, referred for primary PCI.",
                "encounter_type": "emergency",
                "patient": {
                  "age_years": 55,
                  "sex": "male"
                },
                "payer": "bupa",
                "options": {
                  "use_llm": false,
                  "max_candidates": 5,
                  "include_not_coded": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "coded (see coding_validation.valid and clarification_required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeNoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "the request could not be read, or asks for a coding system other than ICD-10-AM",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "a key is required and was not supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "the note is larger than this service accepts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "the request carries a patient identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "the service failed; nothing from the request was stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/code-note/batch": {
      "post": {
        "tags": [
          "code-note"
        ],
        "summary": "Up to 200 notes in one call",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeNoteBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "one result per note, in order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeNoteBatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "the request could not be read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "key required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "too many notes, or too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feedback": {
      "post": {
        "tags": [
          "code-note"
        ],
        "summary": "A coder accepts or rejects a returned code (codes only)",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              },
              "example": {
                "request_id": "6f5dc6bc-1e3c-48dc-a041-8cc9cf4016e9",
                "code": "E11.9",
                "decision": "reject",
                "reason": "more_specific_available",
                "replacement_code": "E11.40"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "400": {
            "description": "unknown field, bad reason or bad request id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "the code is not in the ICD-10-AM table",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "code-note"
        ],
        "summary": "Export feedback for the Coder Review queue and vocabulary review",
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "codes and counts only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackExport"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key", "description": "Only required when the deployment sets RXDX_API_KEY." }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" } } } }
      },
...
    }
  }
}